Add event driven example components
In this example, we are still using promises, due to the nature of
that react DNS resolver. However, now we are firing events whenever
its business is done. This means that other parts of the system can
interact with the results, ... and modern code editors also love to
type hint the objects that are being passed around!